+++ tests/model/test_multigpu.py
@@ -3,6 +3,10 @@ import subprocess
 
 import pytest
 import torch
+from mudata import MuData
+
+import scvi
+from scvi.model import MULTIVI, PEAKVI, TOTALVI, CondSCVI, LinearSCVI
 
 
 @pytest.mark.multigpu
@@ -32,7 +36,6 @@ def test_scanvi_from_scvi_multigpu():
             strategy="ddp_find_unused_parameters_true",
         )
         print("done")
-        torch.distributed.destroy_process_group()
 
         assert model.is_trained
         adata.obsm["scVI"] = model.get_latent_representation()
@@ -61,8 +64,6 @@ def test_scanvi_from_scvi_multigpu():
         )
         print("done")
         adata.obsm["scANVI"] = model_scanvi.get_latent_representation()
-
-        torch.distributed.destroy_process_group()
 
         assert model_scanvi.is_trained
 
@@ -101,13 +102,143 @@ def test_scanvi_from_scratch_multigpu():
         )
         print("done")
 
-        torch.distributed.destroy_process_group()
-
         assert model.is_trained
 
 
 @pytest.mark.multigpu
-def test_scvi_train_ddp(save_path: str = "."):
+def test_totalvi_multigpu():
+    adata = scvi.data.synthetic_iid()
+    protein_adata = scvi.data.synthetic_iid(n_genes=50)
+    mdata = MuData({"rna": adata, "protein": protein_adata})
+    # TOTALVI.setup_anndata(
+    #     adata,
+    #     batch_key="batch",
+    #     protein_expression_obsm_key="protein_expression",
+    #     protein_names_uns_key="protein_names",
+    # )
+    TOTALVI.setup_mudata(
+        mdata,
+        batch_key="batch",
+        modalities={"rna_layer": "rna", "batch_key": "rna", "protein_layer": "protein"},
+    )
+    n_latent = 10
+    model = TOTALVI(mdata, n_latent=n_latent)
+    model.train(
+        100,
+        train_size=0.5,
+        early_stopping=False,
+        check_val_every_n_epoch=1,
+        accelerator="gpu",
+        devices=-1,
+        strategy="ddp_find_unused_parameters_true",
+    )
+    assert model.is_trained is True
+
+
+@pytest.mark.multigpu
+def test_multivi_multigpu():
+    # adata = scvi.data.synthetic_iid()
+    mdata = scvi.data.synthetic_iid(return_mudata=True)
+    # MULTIVI.setup_anndata(
+    #     adata,
+    #     batch_key="batch",
+    # )
+    MULTIVI.setup_mudata(
+        mdata,
+        batch_key="batch",
+        modalities={
+            "rna_layer": "rna",
+            "protein_layer": "protein_expression",
+            "atac_layer": "accessibility",
+        },
+    )
+    n_latent = 10
+    model = MULTIVI(
+        mdata,
+        n_latent=n_latent,
+        n_genes=50,
+        n_regions=50,
+    )
+    model.train(
+        100,
+        train_size=0.5,
+        early_stopping=False,
+        check_val_every_n_epoch=1,
+        accelerator="gpu",
+        devices=-1,
+        strategy="ddp_find_unused_parameters_true",
+    )
+    assert model.is_trained is True
+
+
+@pytest.mark.multigpu
+def test_peakvi_multigpu():
+    adata = scvi.data.synthetic_iid()
+    PEAKVI.setup_anndata(
+        adata,
+        batch_key="batch",
+    )
+
+    model = PEAKVI(
+        adata,
+        model_depth=False,
+    )
+
+    model.train(
+        max_epochs=100,
+        train_size=0.5,
+        check_val_every_n_epoch=1,
+        early_stopping=False,
+        accelerator="gpu",
+        devices=-1,
+        strategy="ddp_find_unused_parameters_true",
+        save_best=False,
+    )
+    assert model.is_trained
+
+
+@pytest.mark.multigpu
+def test_condscvi_multigpu():
+    adata = scvi.data.synthetic_iid()
+    adata.obs["overclustering_vamp"] = list(range(adata.n_obs))
+    CondSCVI.setup_anndata(
+        adata,
+        labels_key="labels",
+    )
+
+    model = CondSCVI(adata)
+
+    model.train(
+        max_epochs=100,
+        train_size=0.9,
+        check_val_every_n_epoch=1,
+        accelerator="gpu",
+        devices=-1,
+        strategy="ddp_find_unused_parameters_true",
+    )
+    assert model.is_trained
+
+
+@pytest.mark.multigpu
+def test_linearcvi_multigpu():
+    adata = scvi.data.synthetic_iid()
+    adata = adata[:, :10].copy()
+    LinearSCVI.setup_anndata(adata)
+    model = LinearSCVI(adata, n_latent=10)
+
+    model.train(
+        max_epochs=100,
+        train_size=0.5,
+        check_val_every_n_epoch=1,
+        accelerator="gpu",
+        devices=-1,
+        strategy="ddp_find_unused_parameters_true",
+    )
+    assert model.is_trained
+
+
+@pytest.mark.multigpu
+def test_scvi_train_ddp(save_path: str):
     training_code = """
 import torch
 import scvi
@@ -125,8 +256,6 @@ model.train(
     devices=-1,
     strategy="ddp_find_unused_parameters_true",
 )
-
-torch.distributed.destroy_process_group()
 
 assert model.is_trained
 """
@@ -160,7 +289,7 @@ assert model.is_trained
 
 
 @pytest.mark.multigpu
-def test_scanvi_train_ddp(save_path: str = "."):
+def test_scanvi_train_ddp(save_path: str):
     training_code = """
 import torch
 import scvi
@@ -189,8 +318,6 @@ model.train(
     strategy="ddp_find_unused_parameters_true",
     datasplitter_kwargs=datasplitter_kwargs,
 )
-
-torch.distributed.destroy_process_group()
 
 assert model.is_trained
 """